Add test suite to help new contributors detect regressions
This is a fairly "plain" test suite based on tox and pytest.  It includes a
linter and code coverage measurement.

The builds use pip-tools to pin requirements in order to make builds
reproducible over time (upstream updates will not cause the build to break).

I tried to make as few changes to the code as I could in order to make it
straightforward to review.  I still made a few changes here and there when it
simplified the test suite.